+++ RELEASES.md
@@ -17,7 +17,7 @@ Upcoming
 - Document JdbcTransactionalStorage type attribute
 - Add storeResultAsStreamInSessionKey to HttpSender
 - Remove IbisLocalSender and IbisJavaSender from pipes package (some time ago deprecated and moved to senders package)
-
+- Add soapAction attribute and parameter to SendTibcoMessage
 
 
 6.0